requirements.txt
langchain==0.0.284
openai==0.27.8
python-dotenv==1.0.0
faiss-cpu==1.7.4
numpy==1.24.3
pandas==2.0.3
streamlit==1.25.0
tiktoken==0.4.0
